♻️ Refactor :  다이나믹 헤더 수정 및 폴더 경로 이동
