+++ src/App.jsx
@@ -5,10 +5,10 @@ import React, { useEffect } from 'react';
 import { Layout } from 'components/layout/Layout';
 import useFetchJobCategories from 'hooks/useFetchJobCategory';
 import NotFound from 'pages/404';
-import CoffeeCreatePage from 'pages/CoffeeCreatePage';
-import CoffeeDetailPage from 'pages/CoffeeDetailPage';
-import CoffeeListPage from 'pages/CoffeeListPage';
-import CoffeeUpdatePage from 'pages/CoffeeUpdatePage';
+import CoffeeCreatePage from 'pages/coffeechats/CoffeeCreatePage';
+import CoffeeDetailPage from 'pages/coffeechats/CoffeeDetailPage';
+import CoffeeListPage from 'pages/coffeechats/CoffeeListPage';
+import CoffeeUpdatePage from 'pages/coffeechats/CoffeeUpdatePage';
 import JdAll from 'pages/jd-all';
 import { JdDetail } from 'pages/jd-detail/JdDetail';
 import { Main } from 'pages/mainpage';
@@ -20,8 +20,8 @@ import MyPage from 'pages/mypage/MyPage';
 import Withdrawal from 'pages/mypage/Withdrawal';
 import RedirectPage from 'pages/sign-in/RedirectPage';
 import SignIn from 'pages/sign-in/SignIn';
-import SignUpFailPage from 'pages/SignUpFailPage';
-import SignUpPage from 'pages/SignUpPage';
+import SignUpFailPage from 'pages/sign-up/SignUpFailPage';
+import SignUpPage from 'pages/sign-up/SignUpPage';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { BrowserRouter, Navigate, Route, Routes } from 'react-router-dom';
 import { useRecoilValue } from 'recoil';
