+++ src/components/coffeechats/detail-coffeechats/ui/DynamicHeader.jsx
@@ -1,11 +1,8 @@
 import Header from 'components/common/header/Header';
-import { COFFEE_CHILD, COFFEE_MYPAGE_CHILD } from 'constants/headerProps';
+import { COFFEE_CHILD } from 'constants/headerProps';
 
 function DynamicHeader({ backPath }) {
-  if (backPath === '/mypage/coffee') {
-    return <Header title={COFFEE_MYPAGE_CHILD.title} url={COFFEE_MYPAGE_CHILD.url} />;
-  } else {
-    return <Header title={COFFEE_CHILD.title} url={COFFEE_CHILD.url} />;
-  }
+  if (backPath) return <Header title={'이전으로'} url={backPath} />;
+  else return <Header title={COFFEE_CHILD.title} url={COFFEE_CHILD.url} />;
 }
 export default DynamicHeader;
